LAI/requirements.txt
# LAI Demo Requirements

--extra-index-url https://download.pytorch.org/whl/cu118
accelerate==0.33.0
bitsandbytes==0.43.3
flake8==4.0.1
flake8-docstrings==1.6.0
mypy==1.8.0
numpy==1.26.4
psutil==6.0.0
PyMuPDF==1.24.10
pytest==7.2.1
pytest-cov==3.0.0
pytest-flake8-v2==1.2.3
pytest-isort==3.0.0
pytest-mypy==0.10.3
pytest-xdist==3.0.2
requests==2.32.3
sentence-transformers==3.0.1
spacy==3.7.6
torch==2.3.1+cu118
tox==3.28.0